.ipynb_checkpoints/requirements-checkpoint.txt
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.0
coloredlogs==15.0.1
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
enum-compat==0.0.3
executing==1.2.0
filelock==3.10.0
flatbuffers==2.0.7
huggingface-hub==0.13.2
humanfriendly==10.0
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.3.0
lit==15.0.7
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.13.1
onnxconverter-common==1.13.0
onnxruntime==1.14.1
onnxruntime-tools==1.7.0
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==11.0.0
pycodestyle==2.10.0
Pygments==2.14.0
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.0.1
regex==2022.10.31
requests==2.28.2
setuptools==65.6.3
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tf2onnx==1.13.0
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torch-model-archiver==0.7.1b20230208
torch-workflow-archiver==0.2.7b20230208
torchserve==0.7.1b20230208
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.38.4
xmltodict==0.13.0
